+++ cron_cashway_check_for_transactions.php
@@ -1,0 +1,37 @@
+<?php
+/**
+ * 2015 CashWay - Epayment Solution
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Academic Free License (AFL 3.0)
+ * that is bundled with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/afl-3.0.php
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@prestashop.com so we can send you a copy immediately.
+ *
+ * DISCLAIMER
+ *
+ * Do not edit or add to this file if you wish to upgrade PrestaShop to newer
+ * versions in the future. If you wish to customize PrestaShop for your
+ * needs please refer to http://www.prestashop.com for more information.
+ *
+ *  @author    hupstream <mailbox@hupstream.com>
+ *  @copyright 2015 Epayment Solution
+ *  @license   http://opensource.org/licenses/afl-3.0.php  Academic Free License (AFL 3.0)
+ */
+
+if (!defined('_PS_ADMIN_DIR_'))
+	define('_PS_ADMIN_DIR_', getcwd());
+
+include(_PS_ADMIN_DIR_.'/../config/config.inc.php');
+include(_PS_MODULE_DIR_.'/cashway/cashway.php');
+
+if (Tools::getIsset('secure_key'))
+{
+	$secureKey = md5(_COOKIE_KEY_.Configuration::get('PS_SHOP_NAME'));
+	if (!empty($secureKey) && $secureKey === Tools::getValue('secure_key'))
+		CashWay::checkForPayments();
+}