+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog for CashWay PrestaShop module
 
+## 18 mars - 0.1.0, dev version
+
+ * added cron task: check local orders pending payment,
+   check transactions status by CashWay service,
+   compare, update.
+
+
 ## 12 mars 2015 - 0.0.1, dev version
 
  * [+] MO : new CashWay module with basic functionality:
    - installs itself with basic config.
    - creates a new transaction on payment method confirmation.
-